+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/deprecated/CamAutoBlueSpecimen.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.autonomous;
+package org.firstinspires.ftc.teamcode.deprecated;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
@@ -7,9 +7,6 @@ import com.qualcomm.robotcore.util.ElapsedTime;
 import org.firstinspires.ftc.teamcode.PathfinderSoftware;
 import org.firstinspires.ftc.teamcode.RobotHardware;
 import org.firstinspires.ftc.teamcode.VisionSoftware;
-
-import java.util.ArrayList;
-import java.util.List;
 
 @Autonomous(name= "BlueSpecimenCam", group ="Blue")
 public class CamAutoBlueSpecimen extends LinearOpMode {
